fixed toLong method of INode

+++ Source/ch/nolix/coreapi/documentapi/nodeapi/INode.java
@@ -130,5 +130,5 @@ public interface INode<N extends INode<N>> extends OptionalHeadered {
 	 * @return the long the current {@link INode} represents.
 	 * @throws RuntimeException if the current {@link INode} does not represent a long.
 	 */
-	int toLong();
+	long toLong();
 }
